Move CF attributes code to dedicated module
Extract CF_ATTRS dictionary and apply_cf_attributes function from
diagnostics.py into a new cf_attributes.py module. diagnostics.py is
intended for routines that compute diagnostic quantities, so CF
metadata handling belongs in its own file. Update fcst.py import
accordingly.
